all register controller fixed

+++ src/utils/cloudinary.js
@@ -1,6 +1,7 @@
 import {v2 as cloudinary} from "cloudinary";
 import fs from "fs";
 
+
 cloudinary.config({
     cloud_name: process.env.CLOUDINARY_CLOUD_NAME,
     api_key: process.env.CLOUDINARY_API_KEY,
@@ -15,18 +16,28 @@ const uploadCloudinary = async (localfilepath) => {
     try{
         if(!localfilepath)
             return null;
-        await cloudinary.uploader.upload(localfilepath, (err,result) => {
-            if(err){
-                console.log(err);
+        const response = await cloudinary.uploader.upload(localfilepath,
+            {
+                resource_type: "auto"
             }
-        })
+        )
         //file has been uploaded to cloudinary
-        console.log("file has been uploaded to cloudinary",response.url);
+        console.log("file has been uploaded to cloudinary",response.secure_url);
+
+        //remove the locally saved file as the upload has failed
+        if (fs.existsSync(localfilepath)) {
+                fs.unlinkSync(localfilepath);
+        }
+
         return response;
     }
-    catch(err){
-        fs.unlinkSync(localfilepath)
-        //remove the locally saved file as the upload has failed
+    catch (err) {
+        console.log("Cloudinary Error:", err);
+
+        if (fs.existsSync(localfilepath)) {
+            fs.unlinkSync(localfilepath);
+        }
+
         return null;
     }
 }
